split curvecamtools from ops.py
removed all calculations which belong in curve cam tools menu from ops.py to their own curvecamtools.py file.
modified the _init_.py to accept the new curvecamtools.py file
removed some imports from top of files
tried to untangle some import *
